fix: scheduler prefill over-commit bug + docs & batched bench
- scheduler.py: prefill fill-in takes remaining slots only
  (min(batch_size - len(running), waiting)); old code could
  over-commit running > batch_size and starve queue tail
- bench_batched.py: continuous batching vs serial benchmark
  (batch 1-4 scan, real wall-clock); batch=4 -> 2.51x throughput
- ENGINE_COMPARISON.md: gap analysis vs industrial engines
  (vLLM) + TODO roadmap (P0-P3)
- PERFORMANCE.md: reorganize by version timeline, move analysis
  to appendices, add v2.0 continuous batching section
- kernels: scaling as tl.constexpr (triton 2.1 compat)
- paged_attention.py: explicit q.float() for mixed-dtype matmul
- batch.py: drop stale Phase 1 docstring

+++ qwen3/paged_attention.py
@@ -99,8 +99,9 @@ def paged_attention(q, k_new, v_new, kv_cache, layer_idx, attention_mask, scalin
         v_page = repeat_kv(v_page, groups).float()
 
         # 部分分数 s^(p) = Q @ K_pageᵀ * scale ∈ R^B（一页的局部得分）
-        s = (torch.matmul(q, k_page.transpose(-2, -1)) * scaling).float()
-        s = s + mask_full[:, :, :, j * page_size:(j + 1) * page_size]
+        # q 需要显式转 float32（bf16 下与 float 的 k_page matmul 会类型报错）
+        s = torch.matmul(q.float(), k_page.transpose(-2, -1)) * scaling
+        s = s + mask_full[:, :, :, j * page_size:(j + 1) * page_size].float()
 
         # online-softmax 合并（对照推导 ①②③④）：
         # ① 新全局最大值 m' = max(m, 本页最大值 m_p)
